+++ src/common.h
@@ -11,7 +11,8 @@
 typedef enum {
     RASPBERRY_PI_PICO = 0,
     WAVESHARE_RP2040_LCD_096,
-    RASPBERRY_PI_PICO_2
+    RASPBERRY_PI_PICO_2,
+    WAVESHARE_RP2350_LCD_096,
 } board_type_t;
 
 typedef enum {