+++ src/main.cpp
@@ -34,7 +34,7 @@ int main() {
     gpio_init(PICO_DEFAULT_LED_PIN);
     gpio_set_dir(PICO_DEFAULT_LED_PIN, GPIO_IN);
 #if defined(RASPBERRYPI_PICO2)
-    board_type_t board_type = RASPBERRY_PI_PICO_2;
+    board_type_t board_type = gpio_get(PICO_DEFAULT_LED_PIN) ? WAVESHARE_RP2350_LCD_096 : RASPBERRY_PI_PICO_2;
 #else
     board_type_t board_type = gpio_get(PICO_DEFAULT_LED_PIN) ? WAVESHARE_RP2040_LCD_096 : RASPBERRY_PI_PICO;
 #endif
@@ -65,6 +65,9 @@ int main() {
             break;
         case RASPBERRY_PI_PICO_2:
             printf("Board: Raspberry Pi Pico 2\r\n");
+            break;
+        case WAVESHARE_RP2350_LCD_096:
+            printf("Board: Waveshare RP2350-LCD-0.96\r\n");
             break;
         default:
             printf("Board: unknown\r\n");